Only compress files that are installed
To avoid failures in the future from changes to the Fedora packages we
use, this patch changes our upx invocation:

* one invocation compresses files we install explicitly
* another lists files from packages we install, and compresses those

In the second phase, missing packages (from changes to dependencies)
will only result in a “package ... is not installed” message on
standard output, which will be filtered by the grep. We ignore errors
because this approach results in asking upx to compress symlinks,
which it complains about (but everything works anyway).

Signed-off-by: Stephen Kitt <skitt@redhat.com>